Refactor voucher discount value to int and cleanups
Changed maximumDiscountValue in VoucherArgument from double to int for consistency. Updated related UI and logic in add/edit voucher screens and cubits to handle int values, converting to double only for display. Added 'unknown' localization key and improved ProductStatusEnum.fromName to handle unknown values. Removed unused edit_product_view.dart.

+++ lib/objects/voucher_related/voucher_argument.dart
@@ -62,7 +62,7 @@ class VoucherArgument {
     bool? isPercentage,
     bool? hasEndTime,
     bool? isLimited,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
     int? maximumUsage,
     int? usageLeft,
     DateTime? endTime,
